+++ src/presentation/screens/index.ts
@@ -23,3 +23,5 @@ export { MuhafizAyarlariSayfasi } from './MuhafizAyarlariSayfasi';
 export { HakkindaSayfasi } from './HakkindaSayfasi';
 export { RamazanAyarlariSayfasi } from './RamazanAyarlariSayfasi';
 export { DebugLogsSayfasi } from './DebugLogsSayfasi';
+export { TakvimAyarlariSayfasi } from './TakvimAyarlariSayfasi';
+export { NelerYeniSayfasi } from './NelerYeniSayfasi';